refactor(emitter): split enum ES5 wrapper
